Reintroduce normalization to isolate mismatches during debugging.
