+++ crates/common/src/fs/mod.rs
@@ -1,3 +1,5 @@
 mod clear_dir;
+mod normalize_path;
 
 pub use clear_dir::*;
+pub use normalize_path::*;
