+++ crates/core/src/helpers/cached_compiler.rs
@@ -185,6 +185,7 @@ async fn compile_contracts(
 
     let compilation = Compiler::new()
         .with_allow_path(&metadata_directory)
+        .with_base_path(metadata_directory.as_ref().to_path_buf())
         // Handling the modes
         .with_optimization(mode.optimize_setting)
         .with_pipeline(mode.pipeline)
